HTML, CSS, & Images
Right now, I've created the the HTML layout which more will be added to the text and other changes will also occur throughout time. The CSS file contains the header and the about section almost finished and we'll be starting going into the characteristics adding color and images. The boxes with text will be changed adding style but after I fill in the necessary information.